+++ changelog.md
@@ -1,5 +1,9 @@
 # ACEmulator Change Log
 
+### 2018-08-10
+[Mag-nus]
+* Significantly improved GetBiota performance from the shard by using flags to indicate populated collections.
+
 ### 2018-08-07
 [Mag-nus]
 * Shard schema changed to remove foreign key link between Character and Biota.